+++ applications/fishing-map/src/features/search/search.slice.ts
@@ -18,7 +18,10 @@ import { getRelatedDatasetByType } from 'features/datasets/datasets.selectors'
 
 export const RESULTS_PER_PAGE = 20
 
-export type VesselWithDatasets = Vessel & { dataset: Dataset; trackDatasetId?: string }
+export type VesselWithDatasets = Omit<Vessel, 'dataset'> & {
+  dataset: Dataset
+  trackDatasetId?: string
+}
 export type SearchType = 'basic' | 'advanced'
 export type SearchFilterKey = 'flags' | 'gearType' | 'startDate' | 'endDate'
 export type SearchFilter = {