+++ applications/fishing-map/src/utils/info.ts
@@ -8,7 +8,8 @@ export const upperFirst = (text: string) =>
   text.charAt(0).toUpperCase() + text.substr(1).toLowerCase()
 
 export const formatInfoField = (fieldValue: string, type: string) => {
-  if (type === 'name') return fieldValue.replace(/\b(?![LXIVCDM]+\b)([A-Z]+)\b/g, upperFirst)
+  if (!fieldValue) return EMPTY_FIELD_PLACEHOLDER
+  if (type === 'name') return fieldValue.replace(/\b(?![LXIVCDM]+\b)([A-Z,Ñ]+)\b/g, upperFirst)
   if (type === 'fleet') {
     const fleetClean = fieldValue.replaceAll('_', ' ')
     return fleetClean.charAt(0).toUpperCase() + fleetClean.slice(1)