+++ packages/api-types/src/vessel.ts
@@ -10,10 +10,13 @@ export interface Vessel {
   shipname: string
   firstTransmissionDate: string
   lastTransmissionDate: string
+  dataset?: string
   imo?: string
   mmsi?: string
   callsign?: string
   fleet?: string
+  casco?: string
+  matricula?: string
   origin?: string
   type?: string
   geartype?: string