+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @globalfishingwatchapp/fishing-map
 
+## 1.1.38
+
+### Patch Changes
+
+- Updated dependencies [72ca7097]
+  - @globalfishingwatch/timebar@2.0.7
+
+## 1.1.37
+
+### Patch Changes
+
+- Updated dependencies [5b01cd38]
+  - @globalfishingwatch/api-types@5.3.0
+
+## 1.1.36
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@9.2.0
+
+## 1.1.35
+
+### Patch Changes
+
+- Updated dependencies [905c8894]
+  - @globalfishingwatch/api-types@5.2.1
+
+## 1.1.34
+
+### Patch Changes
+
+- Updated dependencies [d5c40e3f]
+  - @globalfishingwatch/api-types@5.2.0
+
+## 1.1.33
+
+### Patch Changes
+
+- Updated dependencies [b5bfe9fb]
+  - @globalfishingwatch/ui-components@10.2.0
+
 ## 1.1.32
 
 ### Patch Changes